+++ apps/api-extractor/src/DebugRun.ts
@@ -48,9 +48,11 @@ const extractor: Extractor = new Extractor(
       enabled: true,
       outputFolder: path.join(__dirname, 'debug')
     },
-    packageTypings: {
+    dtsRollup: {
       enabled: true,
-      outputFolder: path.join(__dirname, 'debug')
+      publishFolderForInternal: path.join(__dirname, 'debug/internal'),
+      publishFolderForBeta: path.join(__dirname, 'debug/beta'),
+      publishFolderForPublic: path.join(__dirname, 'debug/public')
     }
   }
 );