+++ apps/rush-lib/src/data/CommandLineConfiguration.ts
@@ -16,6 +16,7 @@ export interface ICustomCommand {
   summary: string;
   documentation: string | undefined;
   parallelized: boolean;
+  ignoreMissingScript: boolean | undefined;
 }
 
 export interface ICustomEnumValue {