+++ common/reviews/api/gulp-core-build.api.ts
@@ -113,6 +113,7 @@ interface IBuildConfig {
   jestEnabled?: boolean;
   libAMDFolder?: string;
   libES6Folder?: string;
+  libESNextFolder?: string;
   libFolder: string;
   onTaskEnd?: (taskName: string, duration: number[], error?: any) => void;
   onTaskStart?: (taskName: string) => void;