+++ core-build/gulp-core-build-mocha/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @microsoft/gulp-core-build-mocha
 
-This log was last generated on Sat, 17 Mar 2018 02:54:22 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Mar 2018 19:12:42 GMT and should not be manually modified.
+
+## 3.3.22
+Mon, 26 Mar 2018 19:12:42 GMT
+
+*Version update only*
+
+## 3.3.21
+Fri, 23 Mar 2018 00:34:53 GMT
+
+*Version update only*
+
+## 3.3.20
+Thu, 22 Mar 2018 18:34:13 GMT
+
+*Version update only*
 
 ## 3.3.19
 Sat, 17 Mar 2018 02:54:22 GMT