+++ core-build/gulp-core-build/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @microsoft/gulp-core-build
 
-This log was last generated on Sat, 17 Mar 2018 02:54:22 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Mar 2018 19:12:42 GMT and should not be manually modified.
+
+## 3.7.2
+Mon, 26 Mar 2018 19:12:42 GMT
+
+### Patches
+
+- Fix an issue where snapshot filenames did not follow Jest's naming convention.
+
+## 3.7.1
+Fri, 23 Mar 2018 00:34:53 GMT
+
+### Patches
+
+- Upgrade colors to version ~1.2.1
+
+## 3.7.0
+Thu, 22 Mar 2018 18:34:13 GMT
+
+### Minor changes
+
+- Add a `libESNextFolder` option
 
 ## 3.6.10
 Sat, 17 Mar 2018 02:54:22 GMT