+++ libraries/load-themed-styles/CHANGELOG.md
@@ -1,6 +1,26 @@
 # Change Log - @microsoft/load-themed-styles
 
-This log was last generated on Tue, 20 Mar 2018 02:44:45 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Mar 2018 19:12:42 GMT and should not be manually modified.
+
+## 1.7.47
+Mon, 26 Mar 2018 19:12:42 GMT
+
+*Version update only*
+
+## 1.7.46
+Sun, 25 Mar 2018 01:26:19 GMT
+
+*Version update only*
+
+## 1.7.45
+Fri, 23 Mar 2018 00:34:53 GMT
+
+*Version update only*
+
+## 1.7.44
+Thu, 22 Mar 2018 18:34:13 GMT
+
+*Version update only*
 
 ## 1.7.43
 Tue, 20 Mar 2018 02:44:45 GMT